backend/requirements.txt
Flask
Flask-CORS
Pillow
requests
python-dotenv
transformers
torch